+++ sdk/resourcemanager/search/armsearch/zz_generated_adminkeys_client.go
@@ -12,13 +12,14 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
+	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 	"net/http"
 	"net/url"
 	"strings"
-
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 )
 
 // AdminKeysClient contains the methods for the AdminKeys group.
@@ -30,8 +31,15 @@ type AdminKeysClient struct {
 }
 
 // NewAdminKeysClient creates a new instance of AdminKeysClient with the specified values.
-func NewAdminKeysClient(con *arm.Connection, subscriptionID string) *AdminKeysClient {
-	return &AdminKeysClient{ep: con.Endpoint(), pl: con.NewPipeline(module, version), subscriptionID: subscriptionID}
+func NewAdminKeysClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) *AdminKeysClient {
+	cp := arm.ClientOptions{}
+	if options != nil {
+		cp = *options
+	}
+	if len(cp.Host) == 0 {
+		cp.Host = arm.AzurePublicCloud
+	}
+	return &AdminKeysClient{subscriptionID: subscriptionID, ep: string(cp.Host), pl: armruntime.NewPipeline(module, version, credential, &cp)}
 }
 
 // Get - Gets the primary and secondary admin API keys for the specified Azure Cognitive Search service.
@@ -84,7 +92,7 @@ func (client *AdminKeysClient) getCreateRequest(ctx context.Context, resourceGro
 func (client *AdminKeysClient) getHandleResponse(resp *http.Response) (AdminKeysGetResponse, error) {
 	result := AdminKeysGetResponse{RawResponse: resp}
 	if err := runtime.UnmarshalAsJSON(resp, &result.AdminKeyResult); err != nil {
-		return AdminKeysGetResponse{}, err
+		return AdminKeysGetResponse{}, runtime.NewResponseError(err, resp)
 	}
 	return result, nil
 }
@@ -156,7 +164,7 @@ func (client *AdminKeysClient) regenerateCreateRequest(ctx context.Context, reso
 func (client *AdminKeysClient) regenerateHandleResponse(resp *http.Response) (AdminKeysRegenerateResponse, error) {
 	result := AdminKeysRegenerateResponse{RawResponse: resp}
 	if err := runtime.UnmarshalAsJSON(resp, &result.AdminKeyResult); err != nil {
-		return AdminKeysRegenerateResponse{}, err
+		return AdminKeysRegenerateResponse{}, runtime.NewResponseError(err, resp)
 	}
 	return result, nil
 }