+++ sdk/resourcemanager/search/armsearch/zz_generated_services_client.go
@@ -12,14 +12,14 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"net/http"
-	"net/url"
-	"strings"
-
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
 	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
+	"net/http"
+	"net/url"
+	"strings"
 )
 
 // ServicesClient contains the methods for the Services group.
@@ -31,8 +31,15 @@ type ServicesClient struct {
 }
 
 // NewServicesClient creates a new instance of ServicesClient with the specified values.
-func NewServicesClient(con *arm.Connection, subscriptionID string) *ServicesClient {
-	return &ServicesClient{ep: con.Endpoint(), pl: con.NewPipeline(module, version), subscriptionID: subscriptionID}
+func NewServicesClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) *ServicesClient {
+	cp := arm.ClientOptions{}
+	if options != nil {
+		cp = *options
+	}
+	if len(cp.Host) == 0 {
+		cp.Host = arm.AzurePublicCloud
+	}
+	return &ServicesClient{subscriptionID: subscriptionID, ep: string(cp.Host), pl: armruntime.NewPipeline(module, version, credential, &cp)}
 }
 
 // CheckNameAvailability - Checks whether or not the given search service name is available for use. Search service names must be globally unique since
@@ -78,7 +85,7 @@ func (client *ServicesClient) checkNameAvailabilityCreateRequest(ctx context.Con
 func (client *ServicesClient) checkNameAvailabilityHandleResponse(resp *http.Response) (ServicesCheckNameAvailabilityResponse, error) {
 	result := ServicesCheckNameAvailabilityResponse{RawResponse: resp}
 	if err := runtime.UnmarshalAsJSON(resp, &result.CheckNameAvailabilityOutput); err != nil {
-		return ServicesCheckNameAvailabilityResponse{}, err
+		return ServicesCheckNameAvailabilityResponse{}, runtime.NewResponseError(err, resp)
 	}
 	return result, nil
 }
@@ -286,7 +293,7 @@ func (client *ServicesClient) getCreateRequest(ctx context.Context, resourceGrou
 func (client *ServicesClient) getHandleResponse(resp *http.Response) (ServicesGetResponse, error) {
 	result := ServicesGetResponse{RawResponse: resp}
 	if err := runtime.UnmarshalAsJSON(resp, &result.SearchService); err != nil {
-		return ServicesGetResponse{}, err
+		return ServicesGetResponse{}, runtime.NewResponseError(err, resp)
 	}
 	return result, nil
 }
@@ -347,7 +354,7 @@ func (client *ServicesClient) listByResourceGroupCreateRequest(ctx context.Conte
 func (client *ServicesClient) listByResourceGroupHandleResponse(resp *http.Response) (ServicesListByResourceGroupResponse, error) {
 	result := ServicesListByResourceGroupResponse{RawResponse: resp}
 	if err := runtime.UnmarshalAsJSON(resp, &result.SearchServiceListResult); err != nil {
-		return ServicesListByResourceGroupResponse{}, err
+		return ServicesListByResourceGroupResponse{}, runtime.NewResponseError(err, resp)
 	}
 	return result, nil
 }
@@ -404,7 +411,7 @@ func (client *ServicesClient) listBySubscriptionCreateRequest(ctx context.Contex
 func (client *ServicesClient) listBySubscriptionHandleResponse(resp *http.Response) (ServicesListBySubscriptionResponse, error) {
 	result := ServicesListBySubscriptionResponse{RawResponse: resp}
 	if err := runtime.UnmarshalAsJSON(resp, &result.SearchServiceListResult); err != nil {
-		return ServicesListBySubscriptionResponse{}, err
+		return ServicesListBySubscriptionResponse{}, runtime.NewResponseError(err, resp)
 	}
 	return result, nil
 }
@@ -472,7 +479,7 @@ func (client *ServicesClient) updateCreateRequest(ctx context.Context, resourceG
 func (client *ServicesClient) updateHandleResponse(resp *http.Response) (ServicesUpdateResponse, error) {
 	result := ServicesUpdateResponse{RawResponse: resp}
 	if err := runtime.UnmarshalAsJSON(resp, &result.SearchService); err != nil {
-		return ServicesUpdateResponse{}, err
+		return ServicesUpdateResponse{}, runtime.NewResponseError(err, resp)
 	}
 	return result, nil
 }