+++ sdk/resourcemanager/search/armsearch/zz_generated_sharedprivatelinkresources_client.go
@@ -12,14 +12,14 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"net/http"
-	"net/url"
-	"strings"
-
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
 	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
+	"net/http"
+	"net/url"
+	"strings"
 )
 
 // SharedPrivateLinkResourcesClient contains the methods for the SharedPrivateLinkResources group.
@@ -31,8 +31,15 @@ type SharedPrivateLinkResourcesClient struct {
 }
 
 // NewSharedPrivateLinkResourcesClient creates a new instance of SharedPrivateLinkResourcesClient with the specified values.
-func NewSharedPrivateLinkResourcesClient(con *arm.Connection, subscriptionID string) *SharedPrivateLinkResourcesClient {
-	return &SharedPrivateLinkResourcesClient{ep: con.Endpoint(), pl: con.NewPipeline(module, version), subscriptionID: subscriptionID}
+func NewSharedPrivateLinkResourcesClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) *SharedPrivateLinkResourcesClient {
+	cp := arm.ClientOptions{}
+	if options != nil {
+		cp = *options
+	}
+	if len(cp.Host) == 0 {
+		cp.Host = arm.AzurePublicCloud
+	}
+	return &SharedPrivateLinkResourcesClient{subscriptionID: subscriptionID, ep: string(cp.Host), pl: armruntime.NewPipeline(module, version, credential, &cp)}
 }
 
 // BeginCreateOrUpdate - Initiates the creation or update of a shared private link resource managed by the search service in the given resource group.
@@ -255,7 +262,7 @@ func (client *SharedPrivateLinkResourcesClient) getCreateRequest(ctx context.Con
 func (client *SharedPrivateLinkResourcesClient) getHandleResponse(resp *http.Response) (SharedPrivateLinkResourcesGetResponse, error) {
 	result := SharedPrivateLinkResourcesGetResponse{RawResponse: resp}
 	if err := runtime.UnmarshalAsJSON(resp, &result.SharedPrivateLinkResource); err != nil {
-		return SharedPrivateLinkResourcesGetResponse{}, err
+		return SharedPrivateLinkResourcesGetResponse{}, runtime.NewResponseError(err, resp)
 	}
 	return result, nil
 }
@@ -320,7 +327,7 @@ func (client *SharedPrivateLinkResourcesClient) listByServiceCreateRequest(ctx c
 func (client *SharedPrivateLinkResourcesClient) listByServiceHandleResponse(resp *http.Response) (SharedPrivateLinkResourcesListByServiceResponse, error) {
 	result := SharedPrivateLinkResourcesListByServiceResponse{RawResponse: resp}
 	if err := runtime.UnmarshalAsJSON(resp, &result.SharedPrivateLinkResourceListResult); err != nil {
-		return SharedPrivateLinkResourcesListByServiceResponse{}, err
+		return SharedPrivateLinkResourcesListByServiceResponse{}, runtime.NewResponseError(err, resp)
 	}
 	return result, nil
 }